Improve handling of XML documents and tests
Use XML documents as input and output of the applier to improve
testability. This ignores the issues of formatted string representations
for the XML which causes assertion problems.

To compare XML documents, the new dependency `xml-unit` has been added
with its `assertj` "flavor" as we use AssertJ everywhere else. The
output for failed tests is very helpful and shows details of the
documents.

The functionally got extended to also find the correct node to modify by
an "eId" attribute.

There was some utility established to help writing more readable code in
the tests. Having whole XML documents in them is quite nice, but easily
gets big. Good formatting is not easy but important.

NDISC-60