Feat(time-machine): ensure target law remains untouched
Kotlin has some "interesting" behavior when it comes to pass-as-value
vs. pass-as-reference for function parameters. Depending on the exact
type complexity, it decides what to choose.
Unfortunately this is not very transparent, nor is it possible to define
read-only parameters. Therefore we need to ensure this desired behavior
with tests.

PS: Cloning a XML document sucks in Java.

NDISC-60