Test(time-machine): simplify XML input strings for test cases
Test cases should be short and concise to what it is about. Therefore
two improvements have been made. First of all, the `asXml` function also
takes care to add the standard header to it can be removed everywhere in
the tests.
Second, it provides some `any*` variables for cases where it simply
doesn't matter and only bloats up the tests and just drags the focus on
what matters for this test. This is a predecessor of the random test
data generator.

NDISC-60