Feat(time-machine): introduce the amending law argument
This removes the layer of the tests providing all arguments of the
modification, but passing an amending law XML to extract this
information from. So far this focuses only on the referenced eId, to
keep the changes small enough.

This includes some "bigger" refactoring in the logical implementation,
just according to TDD. The procedure of the applier logic grows and
demands for structure.

NDISC-60