Feat(system-tests): setup basic system tests with Playwright
As the whole user experience is intended to be within the web-browser,
using an end-to-end test framework for system tests seems like an
obvious solution.

The new folder `system-tests` is intentionally not within any of the
system components as it acts independently on all of them together as
system.

Playwright is a sophisticated framework to test with web-browsers in
user centric way. The team has experience with it and uses it for other
systems of the NeuRIS project too.

NDISC-80