Chore(pipeline): reduce conditional jobs
A few jobs only run on the main branch. This makes sense to not trigger
releases and production deployments on feature branches or dependabot
updates. On the other hand it is the goal to run as much of the original
"main" pipeline on such branches. That increases the trust into merging
these branches and avoids broken pipelines on main after merging.

Therefore it is desirable to make use of SonarCloud its branch/pull
request analysis. This will give early code quality feedback.

The job to release the VSCode extension in turn got this condition.
Before that, it had this condition implicitly via its job dependencies.
As this was resolved, the job itself must have this condition. While it
might be wise to do so explicitly anyhow to avoid such issues.

NDISC-0000