Fix: move icon type declarations to vite env
Apparently having them specified in the tsconfig is enough to make VS
Code happy, but not enough for vue-tsc. For that to work, we would also
need to add the unplugin-icons folder inside node_modules to the
`include` option in tsconfig, which didn't feel good. Adding them to the
environment seems to do the trick for both though. See:

https://github.com/unplugin/unplugin-icons/issues/128

RISDEV-3192