Chore: add patch group to Dependabot
The goal is for Dependabot to create individual PRs for major and minor
version updates, and a grouped PR once a day for patch updates. If this
works I'll also change it for the VS Code extension.

This only affects modules in the NPM ecosystem; other ecosystems have
solved this already.

RISDEV-0000