Create norms and announcements table
This enables us to create our new domain model.

The old tables are not deleted at the moment and the data is not
migrated. We instead create new initial data. We do not think that
saving our current data is worth it.
After we have migrated the backend domain model we can then delete the
tables `amending_law`, `target_law` and `article`

RISDEV-3653