fix: change how layout is implemented
In order to implement some advanced nested scrolling behavior such as
required for the Zeitgrenzen management screen, we need layouting that
is a bit more predictable. Note that this involves making headers
fixed, which is not ideal because it takes up a lot of space. But since
headers are going to be reworked some time soon (hopefully) anyways,
that's fine for now.

RISDEV-3742