Create method to get node by eid
While writing some unit tests using xpaths, I stumbled upon some
limitations of jsdom (the dom implementation used in the tests): It e.g.
does not support upper case characters within either the xpath, which is
something we need for e.g. the eIds (see also:
https://github.com/jsdom/jsdom/issues/2530). Sadly the support for
xpaths is even worse with happy-dom :/

I therefore decided to mock the implementation of the method for
evaluating xpaths in the unit tests using the
 [xpath](https://www.npmjs.com/package/xpath) library.

RISDEV-3932