Add loading states to Affected documents
This required some reworking of the loading structure as before the view
loaded all the nested norms. This is now handled by the nested component
to not need a complicated loop setup and to be able to show the norms
even if some fail to load

This is the first component we have so far that is loading data from the
backend itself. Our histoire does not support fetching data. For now, we
have decided to remove the story for it.

RISDEV-3805