fix: prevent page from breaking when no time boundary is selected
This could happen when-for whatever reason-a law has no time boundaries, or
when fetching time boundaries takes longer than expected. In those cases, links
in the sidebar need to be hidden because they depend on a selected time boundary.
If no time boundary is selected, the router throws a missing param error.

RISDEV-0000