feat: make useElementId easier to use
This commit updates the 'useElementId' composable to return an object that
holds any ID you need on that page automatically. It works by returning a
proxy that checks if a property you're accessing already exists, and if not,
it creates an ID at that property. That way you don't need to call
useElementId multiple times (for each ID you need) anymore. You can call it
once and get an arbitrary number of IDs from the returned object.

Example:

const { id1, id2 } = useElementId()

RISDEV-3831