feat: expose teleport target in header composable
This fixes an issue where the application would crash when trying to teleport
content to the header despite the header already existing in the component
tree. The issue was that while the header might exist in the tree, its HTML
might not have been rendered yet at the time when the child attempts to
teleport content.

To fix this, this commit adds the teleport target to the header context, but
sets it to 'null' until the target is safe to use. Note that the context is
only available to children of the header. Other components can use the slot
of the header.

RISDEV-4188