test: add test for 404 redirect when timeboundary is not found
Currently the metadata page can be opened with any valid date. This should be
changed such that the page ensures that the date is also a valid time boundary,
and redirects to the 404 page if it isn't.

This commit adds E2E tests for the intended behavior. Those are going to fail
for now until implementation is done.

RISDEV-0000