Handle child nodes when replacing text content
This is not 100% correct as now old texts that are split into multiple
child nodes are not detected correctly.

E.g. when replacing "test 123" with "test 789" in
"test <akn:ref>123</akn:ref> 456 test 123"
the correct result would be "test 789 456 test 123" but this
implementation will create "test <akn:ref>123</akn:ref> 456 test 789"

But it doesn't replace all akn:ref elements all the time and is thereby
the better implementation for this prototype

RISDEV-4350