chore: bump RIS UI
This newer version of RIS UI does not have a direct dependency on vue and primevue,
but instead only lists them as peer dependencies. It also is less opinionated about
the specific version of vue and primevue. This should fix some dependency resolution
conflicts that caused the tests to break.

RISDEV-0000