feat: add controller for retrieving norms by their manifestation ELI
I chose to add a new controller for this rather than adding it to the
norms controller because the norm controller currently always expects
expression ELIs. Changing this to support both expression and
manifestation ELIs would mean a bunch of changes to the existing
endpoints, as well as extra boilerplate because we can't use the
controller-level path mapping anymore.

In addition, the manifestation ELI-based endpoints have slightly
different semantics (e.g. the file type will not be provided through a
content header, but by the format property of the ELI).

RISDEV-5107