fix: use generic error message for unhandled exceptions
This commit removes the detailed error output from the fallback exception
handler that is used as controller advice for all exceptions that aren't
handled explicitly. The message is replaced with a generic note that an
error has occurred. If we use the exception message we might leak internal
information such as DB constraints in the API response.

RISDEV-4871