Fix local keycloak setup
We use a service called localhost to be able to address the keycloak
service as "localhost:8443" in the application.yaml of our application.
Using "keycloak:8443" is not possible as spring would then redirect the
user to "keycloak:8443" to login, but that is a host that is not
available on the host system. The spring boot oauth implementation is
very restrictive on the issuer and therefore creates errors if the hosts
used by it and the user to connect to keycloak differ.

See also:
 - https://github.com/spring-projects/spring-security/issues/14633
 - https://github.com/keycloak/keycloak/issues/29783
 - https://github.com/keycloak/keycloak/issues/24252
 - https://medium.com/@kostapchuk/integrating-keycloak-with-spring-boot-in-a-dockerized-environment-813eab1f140c

RISDEV-5805