Add local-only redirect from backend to frontend
When running backend and frontend from different ports, e.g. during
local development, the redirect after the login ends on the backend and
one needs to open the frontend again manually.
This commit adds a controller that redirects to the frontend host for
this case.

We needed to separate the application-local.yaml between running locally
with and without docker as this setup is only needed when running it
without docker and would break the application otherwise.

RISDEV-5805