Fix search path for psql function
This ensures that all calls of the function use the expected searchpath
and the tables of our schema can be found.
As our migration-import doesn't set the searchpath when inserting new
documents but uses fully qualified table names this leads to the job
failing as it is not finding the tables.

RISDEV-6261