Move publish state to norm_manifestation table
To represent this in the code we needed to introduce a new
NormManifestationDto as well as a Repository. This also changed how
norms are loaded in different sports as it is no longer enough to just
load the documents of the given norm. This sadly created some more
changes that needed to be done like adding the GUID of the current
version of norm to the norm_manifestation table as well.
We also moved the loading of the norms for the publishing service to use
the manifestation elis instead of the database ids as the
norm_manifestation table is using those as the primary keys anyway.

In many integration tests we needed to change how the norms are
persisted to the temporary database to just persist the
regelungstext-dokumente as the database will take care of creating the
entries in the norm_manifestation table by itself.

RISDEV-6261