Improve performance of Dokument hashCode
As the Dokumente are stored in HashSets the hashCode is called quite
often now. The previous implementation was really slow as it was parsing
the whole XML-Document to a string. The new implementation hashes only
the fields that Node::isEqualNode is also checking and thereby ~5x
faster for large Dokumente and 9x faster for small once.

RISDEV-6264