Reduce flakiness of integration test
The test failed once as the Instant.now() created for the
afterTwoPublishes was the same as the one used for the lastModifiedAt
date of norm3 as this just happened within the same instant (probably
due to some precision of the Instant). By doing the first publishing of
the changelog in between enough time should have passed that this is no
longer a happening.

RISDEV-0000