fix: allow access to URLs used by the frontend
This fixes an issue where access to any route in the frontend would
always be redirected to "/". The reason was that our security config was
such that Spring Boot would try to authenticate those routes, even
though they should be handled by the frontend and don't exist in the
backend. Because the frontend also handles authentication, the calls
would always fail, since they never reached the frontend.

As a workaround for now, we removed the authentication entry point and
manually listed all top level routes used by the frontend.

RISDEV-6773