test: increase access token lifespan for local Keycloak
This is mostly to simplify E2E testing. Since the E2E tests can
sometimes take more than 5 minutes (the default token lifespan), it
could happen that the token used by the E2E tests expires during
long-running tests. Refreshing the token for the E2E tests is a bit
complicated, so for now we'll just extend the token lifespan of the
local Keycloak to avoid the issue.

RISDEV-0000