Combine public and private bucket UseCases
We are currently working on adjusting the S3Bucket Ports to also support
one additional bucket for the portal prototype. This would mean adding a
3rd port to the once we currently have for publishing public norms and
publishing private norms with exactly the same api and implementation.
To not do that we decided to instead have only one general
PublishNormPort that then uses three different configurations  (public,
private, portal-prototype) that can be chosen by using the @Qualifier
annotation.

In practice this means that we have one PublishNormPort that is
implemented by a BucketService that is not annotated with @Service, and
we have a BucketServiceConfiguration that creates one instance of the
BucketService for each of the 3 configurations and configures those with
the correct S3Clients.

And as an example in the PublishService we then have
    @Qualifier("public")
    PublishNormPort publishPublicNormPort,
    @Qualifier("private")
    PublishNormPort publishPrivateNormPort,
instead of
    PublishPublicNormPort publishPublicNormPort,
    PublishPrivateNormPort publishPrivateNormPort,
The same is also the case for the other bucket specific ports

RISDEV-6988