fix: revert to previous path regex for fallback controller
This reverts the regex used for deciding whether a path should serve an
asset or be redirected to the frontend index.html to an earlier version.
I originally changed it because it doesn't actually work the way I
believe it was intended--it only works if the build output of our
frontend conforms to a very specific format where static assets can only
be at the root or directly inside /assets, but not in any nested
folders. This works for now but can easily break in the future and lead
to hard to debug issues.

Reverting for now to fix the broken pipeline, but this is something we
should keep an eye on.

RISDEV-6804