refactor: separate schema for announcement controller responses
The announcement controller used to share a response schema with the
norm controller, because both returned a norm. Soon, announcements will
contain data that isn't found in the norm (such as the date and time of
the data import). Preparing for this, this commit creates a separate
schema for the announcement controller. Only a clone for now, but will
diverge in the future.

RISDEV-7049