Adjust placement of norms-application-only namespace
It can't be placed directly within the proprietary element as ldml_de
restricts the child elements. We, therefore:
Place the element for this norms-application-only metadata into the
metadata element for RISMetadata
(http://MetadatenRIS.LegalDocML.de/1.7.2/), it is an element for which
we control the schema and are allowed to change it in every way we want.
We could even use two different schema definitions (with and without our
norms-application-only parts) to then validate against the one with our
metadata during the processing and the one without before publishing or
during the import

RISDEV-7104