docs: remove note about process when an announcement is rejected
Since the ZIP file and the contained documents shouldn't change, there
are not plans to have a technical solution in case the announcement is
rejected (note that this doesn't affect re-trying when the upload failed
due to issues unrelated to the content, such as network interruptions).
This means the question about recovery in those cases is outside the
scope of the API specification.

RISDEV-7111