fix: ignore invalid references when returning Zielnormen
The Zielnormen endpoint used to throw a NormNotFoundException, resulting
in a 404 response, if any of the Norms referenced by the metadata could
not be found. This led to confusing behavior in the frontend. Now,
instead of throwing an error, Norms that could not be loaded will simply
be ignored.

RISDEV-6941