fix: remove all sizing and scrolling from preview
So far, the preview used to manage its own scrolling and sizing. This
leads to issues in areas where we want to manage these in the parent
component, e.g. the splitters. Since that is how we're going to be using
this component most of the time going forward, scrolling and sizing are
no longer handled by the preview.

RISDEV-7421