refactor: remove label from date input
This allows us to use the date input as a drop-in replacement for the
InputText component, because they now work in almost the same way. It
also disables all tests in existing components that break due to this
change. The components will be removed soon, so no point in updating
them.

RISDEV-7424