Rework zeitgrenzen updates to keep ids and block updating used once
Previously the ids where updated with every update. This could lead to
changing assigned zeitgrenzen.
Changing this required quite a few changes, so I also extracted the
logic touching the different ldml_de elements from CustomModsMetadata to
individual classes for the different elements.

I also moved the actual update logic into the service as it requires
merging logic that at least to me doesn't really fit into the domain
layer. This was also warrented by changing the usecase to take a custom
data model instead of a Zeitgrenze, which was necessary as the use case
only is supplied with a partial Zeitgrenze and the Zeitgrenzen class now
extracts the information about itself directly from the xml

RISDEV-7669