fix: get Zeitgrenzen colors by reference
We used to get the Zeitgrenze colors based on the eId of the Zeitgrenze.
However the ID of the Zeitgrenze no longer follows eId semantics but is
a UUID, so we can no longer rely on the numbering/sorting to determine
the position of the Zeitgrenze. This Commit changes the logic to resolve
the actual Zeitgrenze based on the ID instead.

Also see: https://github.com/digitalservicebund/ris-norms/pull/1260

RISDEV-7627