Fix orphan removal process in create new expressions
If an expression was created by the in-progress vekuendung, but later the zielnorm-reference is removed, then it is an orphan. In the current inmplementation, where we list all relevant expressions from the start of the first time boundary, it can happens in two ways:
1. Orphan was at the beginning of the timeline of new expressions
2. Orphan is not at the beginning

For the second case, we do it in the loop of the listed expressions. The timeline is fixed automatically (when next expression is created/overriden)

For the first case it won't happen in the loop, since the "new" first time boundary is now later than that expression. Meaning once the loop is done we check the amended-norm-expressions. The fixing of the timeline needs to be done manually here

RISDEV-7181