Simplify EliService::findNextExpressionEli
Instead of searching for a free eli we load the existing elis and
calculate the next free one from that.
This hopefully reduces the db-requests for loading the list for
expressionen-erzeugen quite a bit. (Especially as we do not need to load
the Dokumente and Binary Files for this step)

RISDEV-0000