Add port for updating only norm publish state to speed up publishing
We noticed that the publishing job to the portal buckets takes even
longer since the 1.8.1 update. To figure out why we added some more
logging to it last week and I just looked at them for a bit to figure
out if there are any quick wins to speed it up.
In the last run we spend around 45% of the time updating the publish
state in our DB and around the same amount of time for the actual
uploading to the two S3-buckets. The remaining 10% where mostly spend
loading the norm from the DB. The other processing we are doing is
pretty much negligible.
The updating of the publish state takes so long as we are saving the
whole norm. Therefore, we added a new port for only updating the publish
state and use that instead.
This should speed up the PublishService by almost 2x

RISDEV-0000