Load all norm work list data in one db-request
Before we loaded all dokumente of all the norms we wanted to load into
memory to then run the xpaths to find the titles. This changes it such
that the sql query directly queries these information from the db.

Old time for loading a page with 100 norms on staging ~6s.

For further speed up we could precompute the title and cache it in the
database (e.g. using a "generated always as" column) This should speed
the query (according to my testing) by a further 2-3x. But this would
increase the time our import takes, as the titles would then be
calculated during that. So if this increases the performance enough I
would like to avoid that

RISDEV-8447