Replace ZIP entry size check
Before we were using getSize() for checking the size of the entry because it is an optional metadata and not guaranteed to be accurate or present. Malicious ZIP files can spoof this value to evade size checks (e.g. in a zip bomb). By manually counting the uncompressed bytes read from the stream, we enforce accurate size validation and improve protection against oversized or malformed inputs.

RISDEV-8326