Remove toHaveBeenCalledWith
New eslint rule to replace toHaveBeenCalledWith with toHaveBeenCalledExactlyOnceWith.

This is only possible if in the test there was only 1 call, otherwise I introduced the nth-called method

For the useToast tests the addToast was being called twice when it was not called with the success severity. First with the info severity and then with the success/fail severity, that is why I changed the toHaveBeenCalledWith to
toHaveBeenLastCalledWith but not sure why is this happening, and if this is expected

For the use404Redirect the toHaveBeenCalledOnce had to be replaced with toHaveBeenCalledExactlyOnceWith too and for that the parameter had to be passed.

RISDEV-0000